refactor search filters
