+++ src/app/search/search-filters/skill-filter.tsx
@@ -1,40 +1,40 @@
 import { useFormContext } from "@/lib/providers/form-provider"
 import { FormLabel } from "@mui/material"
-import { DutyDto } from "../api/duties/handler"
-import MultiSelect from "./multi-select"
+import { SkillDto } from "../../api/skills/handler"
+import MultiSelect from "../multi-select"
 
-export interface DutyFilterProps {
-    duties: DutyDto[]
+export interface SkillFilterProps {
+    skills: SkillDto[]
 }
 
-export function DutyFilter({ duties }: DutyFilterProps) {
+export function SkillFilter({ skills }: SkillFilterProps) {
     const { form } = useFormContext()
-    const included = form.watch("duties.include")
-    const excluded = form.watch("duties.exclude")
+    const included = form.watch("skills.include")
+    const excluded = form.watch("skills.exclude")
 
     return (
         <section>
             <div className="pb-4">
-                <FormLabel>Responsibilities</FormLabel>
+                <FormLabel>Skills</FormLabel>
             </div>
 
             <div className="flex flex-col gap-4">
                 <MultiSelect
                     form={form}
-                    controlName="duties.include"
-                    options={duties}
+                    controlName="skills.include"
+                    options={skills}
                     disabledOptions={excluded.map((v) => v.id)}
                     label="Include"
-                    ariaLabel="Duties Included"
+                    ariaLabel="Skills Included"
                 />
 
                 <MultiSelect
                     form={form}
-                    controlName="duties.exclude"
-                    options={duties}
+                    controlName="skills.exclude"
+                    options={skills}
                     disabledOptions={included.map((v) => v.id)}
                     label="Exclude"
-                    ariaLabel="Duties Excluded"
+                    ariaLabel="Skills Excluded"
                 />
             </div>
         </section>
