--- src/app/search/search.module.scss
@@ -1,24 +1,0 @@
-.search-form {
-    @apply min-h-0 flex flex-col;
-
-    hr {
-        @apply my-2;
-    }
-
-    h1,
-    h2 {
-        @apply leading-tight;
-    }
-
-    h1 {
-        @apply mb-2;
-    }
-
-    h2 {
-        @apply mb-1;
-    }
-}
-
-.instructions {
-    @apply text-xs italic mt-2;
-}
