+++ src/app/search/search.tsx
@@ -1,25 +1,18 @@
-import { useDutyDto, useLocationDto, useSkillDto } from "@/lib/hooks/api"
 import { useFormContext } from "@/lib/providers/form-provider"
-import { Button, Divider, Paper, TextField } from "@mui/material"
-import { ClearanceFilter } from "./clearance-filter"
-import { DutyFilter } from "./duty-filter"
-import { LocationFilter } from "./location-filter"
-import { LocationTypeFilter } from "./location-type-filter"
-import styles from "./search.module.scss"
-import { SkillFilter } from "./skill-filter"
-import { YoeFilter } from "./yoe-filter"
+import { Button, Paper } from "@mui/material"
+import { SearchFilters } from "./search-filters/search-filters"
 
 export interface SearchProps {
     className?: string
 }
 
 export default function Search({ className }: SearchProps) {
-    const { form, handleSubmit, handleReset, handleClear } = useFormContext()
-    const { register, formState } = form
-
-    const { data: duties } = useDutyDto()
-    const { data: skills } = useSkillDto()
-    const { data: locations } = useLocationDto()
+    const {
+        form: { formState },
+        handleSubmit,
+        handleReset,
+        handleClear,
+    } = useFormContext()
 
     return (
         <form
@@ -27,70 +20,13 @@ export default function Search({ className }: SearchProps) {
                 ev.preventDefault()
                 handleSubmit()
             }}
-            className={`${styles["search-form"]} ${className ?? ""}`}
+            className={`min-h-0 flex flex-col ${className ?? ""}`}
         >
             <Paper
                 variant="outlined"
                 className="overflow-auto flex flex-col p-2"
             >
-                {/* Text filter */}
-                <div className="px-2 pt-2">
-                    <TextField
-                        label="Text"
-                        variant="standard"
-                        type="text"
-                        placeholder="software (dev|eng).*"
-                        {...register("text")}
-                    />
-                </div>
-
-                <div className="pb-2 pt-4">
-                    <LocationFilter locations={locations ?? []} />
-                </div>
-
-                <div className="pb-2 pt-4">
-                    <Divider />
-                </div>
-
-                {/* Experience filter */}
-                <div className="px-2">
-                    <YoeFilter />
-                </div>
-
-                <div className="pb-2">
-                    <Divider />
-                </div>
-
-                {/* Skill / duty filters */}
-                <div className="px-2">
-                    <SkillFilter skills={skills ?? []} />
-                </div>
-
-                <div className="pb-2 pt-4">
-                    <Divider />
-                </div>
-
-                <div className="px-2">
-                    <DutyFilter duties={duties ?? []} />
-                </div>
-
-                <div className="pb-2 pt-4">
-                    <Divider />
-                </div>
-
-                {/* Miscellaenous */}
-                <section className="flex flex-col px-2">
-                    <div className="flex flex-col gap-4">
-                        <TextField
-                            label="Salary"
-                            variant="standard"
-                            type="number"
-                            {...register("salary")}
-                        />
-                        <LocationTypeFilter />
-                        <ClearanceFilter />
-                    </div>
-                </section>
+                <SearchFilters />
             </Paper>
 
             <div className="pt-4 flex justify-end gap-2">
